refactor: centralize spinner and empty state
