+++ app/(tabs)/index.tsx
@@ -29,8 +29,8 @@ import { useTheme } from '@/contexts/ThemeContext';
 import HomeHeader from '@/features/home/components/HomeHeader';
 import CategoryTabs from '@/features/home/components/CategoryTabs';
 import ProductGrid from '@/features/home/components/ProductGrid';
-import Spinner from '@/features/home/components/Spinner';
-import EmptyState from '@/features/home/components/EmptyState';
+import Spinner from '@/shared/ui/Spinner';
+import EmptyState from '@/shared/ui/EmptyState';
 import ErrorBoundary from '@/components/ui/ErrorBoundary';
 import BannerFormModal from '@/components/BannerFormModal';
 import CartModal from '@/features/cart/components/CartModal';
