+++ app/(tabs)/notifications.tsx
@@ -6,8 +6,8 @@ import { useAuth } from '@/features/auth/AuthContext';
 import { useTheme } from '../../contexts/ThemeContext';
 import { useLanguage } from '../../contexts/LanguageContext';
 import AppShell from '../../components/layout/AppShell';
-import Spinner from '../../components/ui/Spinner';
-import EmptyState from '../../components/ui/EmptyState';
+import Spinner from '@/shared/ui/Spinner';
+import EmptyState from '@/shared/ui/EmptyState';
 import InfoModal from '../../components/InfoModal';
 import { useAuthModal } from '@/features/auth/AuthModalContext';
 import { useNotifications } from '../../src/features/notifications/hooks/useNotifications';
