+++ app/store/[storeId]/admin/_PricingTiersScreen.tsx
@@ -23,7 +23,7 @@ import { useTheme } from '../../../../contexts/ThemeContext';
 import { useCurrency } from '../../../../contexts/CurrencyContext';
 import DatabaseService from '../../../../services/database';
 import { PricingTier } from '../../../../types';
-import Spinner from '../../../../components/ui/Spinner';
+import Spinner from '@/shared/ui/Spinner';
 import InfoModal from '../../../../components/InfoModal';
 import PricingTierFormModal from '@/features/products/components/PricingTierFormModal';
 import Card from '../../../../components/Card';
