+++ app/store/[storeId]/admin/pricing-tiers.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '../../../../components/ui/Spinner';
+import Spinner from '@/shared/ui/Spinner';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const PricingTiersScreen = React.lazy(() => import('./_PricingTiersScreen'));
