+++ app/store/[storeId]/admin/settings.tsx
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import Spinner from '../../../../components/ui/Spinner';
+import Spinner from '@/shared/ui/Spinner';
 import RequireWallet from '../../../../components/RequireWallet';
 
 const SettingsScreen = React.lazy(() => import('./_SettingsScreen'));
