+++ src/shared/ui/EmptyState.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 import type { LucideProps } from 'lucide-react-native';
-import { useTheme } from '../../contexts/ThemeContext';
-import Button from './Button';
+import { useTheme } from '@/contexts/ThemeContext';
+import Button from '@/components/ui/Button';
 
 type IconComponent = React.ComponentType<LucideProps>;
