+++ src/shared/ui/Spinner.tsx
@@ -7,7 +7,7 @@ import {
   StyleProp,
   ViewStyle,
 } from 'react-native';
-import { useTheme } from '../../contexts/ThemeContext';
+import { useTheme } from '@/contexts/ThemeContext';
 import { spacing } from '@/constants/tokens';
 
 interface SpinnerProps {
